+++ src/base/AddressZoneInterface.php
@@ -60,6 +60,7 @@ interface AddressZoneInterface
      * Return the zip code match
      *
      * @return string
+     * @since 2.2
      */
     public function getZipCodeConditionFormula(): string;
 }