+++ src/models/TaxAddressZone.php
@@ -56,6 +56,7 @@ class TaxAddressZone extends Model implements AddressZoneInterface
 
     /**
      * @var string The code to match the zip code.
+     * @since 2.2
      */
     public $zipCodeConditionFormula;
 
@@ -192,6 +193,7 @@ class TaxAddressZone extends Model implements AddressZoneInterface
 
     /**
      * @return string
+     * @since 2.2
      */
     public function getZipCodeConditionFormula(): string
     {