+++ src/services/Subscriptions.php
@@ -407,10 +407,8 @@ class Subscriptions extends Component
         }
 
         $response = $gateway->subscribe($user, $plan, $parameters);
-
-        if ($response->isInactive()) {
-            throw new SubscriptionException(Plugin::t( 'Unable to subscribe at this time.'));
-        }
+        
+        $failedToStart = $response->isInactive();
 
         $subscription = new Subscription();
         $subscription->userId = $user->id;
@@ -423,6 +421,13 @@ class Subscriptions extends Component
         $subscription->subscriptionData = $response->getData();
         $subscription->isCanceled = false;
         $subscription->isExpired = false;
+        $subscription->hasStarted = !$failedToStart;
+        $subscription->isSuspended = $failedToStart;
+
+        if ($failedToStart) {
+            $subscription->dateSuspended = Db::prepareDateForDb(new DateTime());
+        }
+
         $subscription->setFieldValues($fieldValues);
 
         Craft::$app->getElements()->saveElement($subscription, false);